ajustes1
alinhamento de texto
