+++ src/pages/Dashboard/Dashboard.css
@@ -129,9 +129,10 @@
 }
 
 .dashboard-user-info .row {
-  align-items: center;
+  align-items: flex-start;
   justify-content: flex-start;
   text-align: left;
+
 }
 
 .dashboard-user-info strong {
@@ -158,7 +159,7 @@
 /* Garantir alinhamento consistente dos labels e valores */
 .dashboard-user-info .row .row > [class*="col"] {
   display: flex;
-  align-items: center;
+  align-items: flex-start;
   min-height: 48px;
   justify-content: flex-start;
   text-align: left;
@@ -172,12 +173,13 @@
 /* Ajuste específico para a linha de endereço (Bairro, Cidade, UF) */
 .dashboard-user-info .row:last-child {
   margin-bottom: 0;
-   justify-content: flex-start;
+  justify-content: flex-start;
+  align-items: flex-start;
   text-align: left;
 }
 
 .dashboard-user-info .row:last-child .row {
-  align-items: center;
+  align-items: flex-start;
   min-height: 48px;
   justify-content: flex-start;
   text-align: left;
@@ -196,7 +198,7 @@
 .dashboard-user-info .row:last-child > [class*="col"] {
   min-height: 48px;
   display: flex;
-  align-items: center;
+  align-items: flex-start;
   justify-content: flex-start;
   text-align: left;
 }
